pyproject.toml
[project]
name = "vegeta"
version = "0.1.0"
description = "Value Estimation Gain Evaluation Tool & Analysis (VEGETA)"
readme = "README.md"
requires-python = ">=3.11"
authors = [
  { name = "VEGETA Team" }
]

dependencies = [
  "typer>=0.12.5",
  "rich>=13.7.1",
  "pyyaml>=6.0.1",
]

[project.optional-dependencies]
dev = [
  "ruff>=0.4.10",
  "mypy>=1.10.0",
  "pytest>=8.2.0",
  "typer[all]>=0.12.5",
]

[project.scripts]
vegeta = "cli.decide:app"

[build-system]
requires = ["setuptools>=68", "wheel"]
build-backend = "setuptools.build_meta"

[tool.ruff]
line-length = 100
target-version = "py311"

[tool.mypy]
python_version = "3.11"
warn_unused_ignores = true
warn_redundant_casts = true
warn_unreachable = true
disallow_untyped_defs = true
no_implicit_optional = true

[tool.setuptools]
packages = [
  "cli",
  "decider",
  "ingest",
  "retrieval",
  "schema",
  "domain_packs",
  "config",
  "db",
  "eval",
  "features",
  "policy",
  "env",
  "agent",
  "llm",
]

[tool.setuptools.package-data]
domain_packs = ["**/*.yaml"]


